pyproject.toml
[build-system]
requires = ["setuptools>=78.1.0"]
build-backend = "setuptools.build_meta"

[project]
name = "frankstate"
version = "0.1.1"
description = "Frank brings order and helps you build complex LLM workflows using scalable, testable, and reusable components."
license = "MIT"
readme = { file = "README-pypi.md", content-type = "text/markdown" }
requires-python = ">=3.12.3"
dependencies = [
    "langchain-core>=1.3.0,<1.4",
    "langgraph>=1.1.8,<1.2",
    "pydantic>=2.12.5,<3",
]
authors = [
    { name = "aamaragones", email = "aangel.maragones@outlook.com" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"
]
keywords = ["LLM", "azure", "langgraph", "langchain", "azure-foundry", "unstructured", "multimodal"]

[project.optional-dependencies]
examples = [
    "langchain",
    "PyYAML",
    "requests",
    "azure-functions",
    "azure-keyvault-secrets",
    "azure-identity",
    "azure-search-documents",
    "azure-storage-blob",
    "langchain-azure-ai[opentelemetry]",
    "langchain-ollama",
    "langchain-chroma",
    "langchain-classic",
    "langchain-mcp-adapters",
    "fastmcp",
    "Pillow",
    "tenacity",
    "unstructured[pdf]",
    "ipykernel",
    "ipywidgets",
    "matplotlib",
]

[project.urls]
homepage = "https://github.com/eiingeeel/frankenst-ai"
changelog = "https://github.com/eiingeeel/frankenst-ai/blob/main/CHANGELOG.md"
issues = "https://github.com/eiingeeel/frankenst-ai/issues"
repository = "https://github.com/eiingeeel/frankenst-ai"

[dependency-groups]
lint = [
    "mypy==1.18.2",
    "ruff==0.11.13",
]
test = [
    "pytest==9.0.2",
    "pytest-cov==7.0.0",
]
dev = [
    { include-group = "lint" },
    { include-group = "test" },
    "build==1.4.2",
    "twine==6.2.0",
]

[tool.uv]
default-groups = ["dev"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["frankstate", "frankstate.*"]
namespaces = false
# The published core wheel intentionally contains only `frankstate`.
# `core_examples` remains the documented reference package of this repository.

[tool.ruff]
line-length = 88
indent-width = 4
extend-include = ["*.ipynb"]
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
files = ["src/frankstate"]
explicit_package_bases = true
mypy_path = ["src"]
warn_no_return = false
warn_unused_ignores = true
warn_redundant_casts = true
allow_redefinition = true

[tool.pytest.ini_options]
testpaths = ["tests/unit_test", "tests/integration_test"]
pythonpath = ["."]
python_files = ["test_*.py"]
addopts = "-ra"
# `tests/unit_test/frankstate` contains the Frankstate patterns suite.
# `tests/unit_test/core_examples` contains example/reference tests for layouts and
# runnable builders built on top of the Frankenst-AI patterns.
# `tests/integration_test` contains smoke coverage for compiled workflows and
# packaging-adjacent execution flows that should run in the default suite.
markers = [
    "unit: fast Frankenst-AI contract tests",
    "integration: compiled workflows and external services (e.g., FastAPI , Microsoft Foundry)",
]